finish renaming map package to map
