+++ lexers/regex/verifiedlexer/src/main/scala/ch/epfl/lexer/OptimisedChecks.scala
@@ -1,7 +1,7 @@
 /** Author: Samuel Chassot
   */
 
-package ch.epfl.chassot
+package ch.epfl.lexer
 
 object OptimisedChecks {
   extension [T](inline value: T) inline def.ensuring(condition: T => Boolean): T = value
